refactor: extract event into zolana-event crate
Move event.rs, the tag module, and the OutputUtxo struct into a new
program-libs/event crate. zolana-interface depends on and re-exports all
three at their original paths (event, instruction::tag,
instruction::OutputUtxo), so the program and all consumers are unchanged.
The new crate has no dependency on zolana-interface, avoiding a cycle.
