+++ program-libs/event/src/lib.rs
@@ -1,7 +1,10 @@
+pub mod output_utxo;
+pub mod tag;
+
+pub use output_utxo::OutputUtxo;
+
 use borsh::{BorshDeserialize, BorshSerialize};
 use solana_pubkey::Pubkey;
-
-use crate::instruction::{tag, OutputUtxo};
 
 /// `GeneralEvent`, emitted via the `emit_event` self-CPI by state-changing
 /// instructions (spec: General Event). It records the queue sequence numbers and
